changed confirmcases route to a component and rearranged nearest hospital structure
