+++ client/src/components/Input/SearchBar.js
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from "styled-components";
 import { SearchBar } from 'antd-mobile';
 
-import { theme } from "../../constants/theme";
+import { theme, mq } from "../../constants/theme";
 const { colors } = theme;
 
 
@@ -12,6 +12,11 @@ const StyledSearchBar = styled(SearchBar)`
     background-color: ${colors.white};
     border: 0.1rem solid rgba(0, 0, 0, 0.5);
     border-radius: 4rem;
+    @media screen and (max-width: ${mq.phone.wide.maxWidth}) {
+      border-radius: 0;
+      border: 0;
+      border-bottom: 0.1rem solid rgba(0, 0, 0, 0.5);
+    }
   }
   .am-search-cancel-show {
     display: none;
