+++ client/src/templates/layouts/MobileTabs.js
@@ -1,6 +1,6 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import styled from "styled-components";
-
+import LocalEmergencyNumber from "../../components/NearestHospital/LocalEmergencyNumber";
 import { Tabs } from 'antd-mobile';
 import { NavLink } from 'react-router-dom';
 
@@ -42,9 +42,29 @@ const MobileTabs = props => {
 
   const { tabIndex, childComponent } = props;
 
+  const [ status, setStatus ] = useState(true);
+
+  useEffect(() => {
+
+    const noticeBar = sessionStorage.getItem("LocalEmergencyBox");
+    if(noticeBar === "true") {
+      setStatus(true);
+    } else if(noticeBar === "false") {
+      setStatus(false)
+    } else {
+      sessionStorage.setItem("LocalEmergencyBox", "true");
+    }
+    console.log(status)
+  }, [])
+
+  const removeNoticeBar = () => {
+    sessionStorage.setItem("LocalEmergencyBox", "false");
+  }
+
 
   return (
     <TabsContainer>
+     {status ? <LocalEmergencyNumber onClick={removeNoticeBar} /> : null}
       <Tabs
         tabs={tabs}
         page={tabIndex}
